server/requirements.txt
Django==5.1.1
django-allauth==65.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
oauthlib==3.2.2
pycparser==2.22
django-cors-headers==4.0.0
load_dotenv==0.1.3
